Add method to get PSN account ID
PSN accounts have an account ID which is different to the account's
username. We can get the account ID using the PSN API package. This adds
a method that wraps around the relevant PSN API method

This will only work for exact matches. If searching for `koala`, an
account with the username `koala` will appear first in the results
returned by the PSN API method, but other accounts with `koala` in the
username will also appear. This method filters the results to return
only the ID for an account with precisely `koala` as the username

In the case that there's no match, this will currently return
`undefined` - this may need revising later